Fix bug with preview showing too many exercises
Previously there was no removing of exercises from selected_exercises_
in mainwindow. This created a situation where any exercises that has
even been into the routine would be in the preview. Regardless of
whether it was currently in it. With this change, the mainwindow now
gets the routine exercises directly from the routine layout instead of
having its own container.